fix: fix windows

+++ tests/core/algorithm/cluster/opt_kmeans_cluster_test.cc
@@ -230,7 +230,7 @@ TEST(OptKmeansCluster, IN4General) {
 
   std::random_device rd;
   std::mt19937 gen(rd());
-  std::uniform_int_distribution<uint8_t> dist(0, UINT8_MAX);
+  std::uniform_int_distribution<unsigned short> dist(0, UINT8_MAX);
 
   for (uint32_t i = 0; i < count; ++i) {
     std::vector<uint8_t> vec(dimension / 2);
